ImportanceSampling
To use importance Sampling in order to use quasi monte carlo to
approximate the expectation when the log return is assumed to be log
normally distributed. The idea is to use student t as a seed. We
simulate student t distributed random variables and then we approximate
the expectation under t. This can be done by multiplying the pdf of t /
pdf of t to the integrand. See report 1.